feat: add contracts routes for operations on all contracts
feat: add basic auth for non get routes

The contracts routes will be used to perform operations that cover more than just a single type of contract.
The first implementation is to ensure when a serviceoffering is removed from the catalog, all contracts that have that service offering are also altered (or removed for bilateral ones).